# docs/reference/openapi.yaml
...
openapi: 3.1.0
info:
  title: Narad HTTP API
  version: v1
  summary: Produce, consume and manage topics on a Narad cluster over HTTP.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  x-narad-lead: |
    Look up every Narad HTTP endpoint: its parameters, request body, the
    grant it needs, the status codes it returns and a real example.
  x-narad-lead-example:
    request: |
      curl -i -u "$AUTH" "$NARAD/v1/topics/orders?partition=1"
    response: |
      HTTP/1.1 200 OK
      Content-Length: 456
      Content-Type: application/json
      Date: Mon, 28 Sep 2026 19:31:43 GMT

      {
        "name": "orders",
        "id": "681a429c0b683b2d",
        "partitions": 3,
        "retention_ms": 172800000,
        "visibility_timeout_ms": 30000,
        "max_in_flight_per_partition": 1024,
        "max_acked_ahead_per_partition": 1024,
        "created_at": 1790623903,
        "owner": "admin",
        "role": "parent",
        "children": ["orders-audit"],
        "schema_version": 0,
        "partition_stats": [
          {
            "index": 1,
            "segments": 1,
            "oldest_offset": 0,
            "next_offset": 2,
            "high_watermark": 2,
            "size_bytes": 166,
            "oldest_segment_at": 1790623903,
            "owner_node": "narad-0"
          }
        ]
      }
  x-narad-placeholders: |
    - `$NARAD` is the base URL of any node or of the load balancer in front
...
      off and `-u "$AUTH"` can be left out.

...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - basicAuth: []
...
paths:
  /v1/topics:
    post:
      operationId: createTopic
      tags: [Topics]
      summary: Create a topic
      description: |
        Creates a topic and, with `parent`, links it as a fan-out child of
        an existing topic in the same call. The caller becomes the topic's
        [owner](glossary.md#owner). A field left out, or sent as `0`, takes
        the server default, except `retention_ms`, where `0` keeps messages
        forever and only leaving it out takes the default.
      x-narad-grant: |
        `create` on the topic name. With `parent`, also ownership of the
        parent or `admin`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTopicRequest'
      responses:
        '201':
          description: Created. The body is the new topic.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Topic'
        '400':
          description: A field is invalid, the name is not allowed, or the schema cannot be registered (from v3.1.0, that includes a schema whose validation would cost too much).
        '401':
          description: Missing or wrong credentials.
        '403':
          description: No `create` grant on the name, or no right to manage `parent`, as the node that answers or the cluster leader sees it.
        '404':
          description: '`parent` does not exist, also after the answering node caught up with the leader.'
        '409':
          description: The topic exists, a topic exists whose name differs only in letter case (the error names it), `parent` cannot take this child (role, child limit, schema, or a delay the parent's retention cannot hold), or (from v3.1.0) the schema, or the parent's schema history a child adopts, would pass a schema byte budget, or `parent` was recreated under the request twice in a row (`topic changed since it was read`).
        '413':
          description: The body is over 1 MiB.
        '415':
          description: No accepted `Content-Type` and no `X-Narad-Client` header.
        '503':
          description: The cluster has no leader to write the topic, the answering node could not reach the leader to confirm `parent`, or (from v3.1.0) every live node is being decommissioned, so no node can take the new partitions; nothing was written.
      x-narad-examples:
        - request: |
            curl -i -u "$AUTH" -X POST "$NARAD/v1/topics" \
              -H "Content-Type: application/json" \
              -d '{"name": "orders", "partitions": 3}'
          response: |
            HTTP/1.1 201 Created
            Content-Length: 224
            Content-Type: application/json
            Date: Mon, 28 Sep 2026 19:31:43 GMT

            {"name":"orders","id":"681a429c0b683b2d","partitions":3,"retention_ms":604800000,"visibility_timeout_ms":30000,"max_in_flight_per_partition":1024,"max_acked_ahead_per_partition":1024,"created_at":1790623903,"owner":"admin"}
    get:
      operationId: listTopics
      tags: [Topics]
      summary: List topics
      description: |
        Lists topics in name order, one page at a time. See
        [Pagination](#pagination).
      x-narad-grant: |
        Any user. The list holds only topics the caller could read with
        [get a topic](#get-topic).
      parameters:
        - name: limit
          in: query
          description: Page size. Values above 1000 are treated as 1000.
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: page_token
          in: query
          description: The `next_page_token` of the previous page. Leave it out for the first page.
          schema:
            type: string
      responses:
        '200':
          description: A page of topics.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicPage'
        '400':
          description: '`limit` is not a positive integer.'
        '401':
          description: Missing or wrong credentials.
        '500':
          description: The node could not read its topic list.
      x-narad-examples:
        - request: |
            curl -i -u "$AUTH" "$NARAD/v1/topics?limit=2"
          response: |
            HTTP/1.1 200 OK
            Content-Length: 540
            Content-Type: application/json
            Date: Mon, 28 Sep 2026 19:31:43 GMT

            {"next_page_token":"orders-audit","topics":[{"name":"orders","id":"681a429c0b683b2d","partitions":3,"retention_ms":604800000,"visibility_timeout_ms":30000,"max_in_flight_per_partition":1024,"max_acked_ahead_per_partition":1024,"created_at":1790623903,"owner":"admin","role":"standalone"},{"name":"orders-audit","id":"91ad229a6dd04bcf","partitions":3,"retention_ms":604800000,"visibility_timeout_ms":30000,"max_in_flight_per_partition":1024,"max_acked_ahead_per_partition":1024,"created_at":1790623903,"owner":"admin","role":"standalone"}]}
  /v1/topics/{topic}:
    parameters:
      - $ref: '#/components/parameters/TopicPath'
    get:
      operationId: getTopic
      tags: [Topics]
      summary: Get a topic
      description: |
        Returns the topic, its current schema and statistics for each
        partition. The node that answers asks every partition's owner for
        its statistics.

        **New in v3.1.0:** when some partitions' owners are down, the answer
        is still `200`. The partitions that could be read carry their
        statistics and `status` `ok`; each other one is a placeholder with
        zero statistics, `status` `owner_unavailable` and the owner's
        `owner_liveness`, and the body carries `partial: true`. Leave the
        placeholders out of any total. An owner that does not answer within
        2 seconds counts as `unreachable`. A v3.0.1 node answers `421`
        instead.

        **Unreleased:** a [remote child](glossary.md#remote-child)'s stub
        has `partitions: 0` and a `remote` object that names the remote and
        the topic there. Its `created_by` and `paused_by` are shown to
        admins only.
      x-narad-grant: |
        Any grant that matches the topic name, ownership, or `admin`.
      parameters:
        - name: partition
          in: query
          description: Return statistics for this partition only.
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: The topic with partition statistics.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicDetails'
        '400':
          description: '`partition` is not a partition of the topic.'
        '401':
          description: Missing or wrong credentials.
        '403':
          description: No grant on the topic.
        '404':
          description: The topic does not exist.
        '421':
          description: From a v3.0.1 node, the owner of one of the topic's partitions could not be found or refused to answer. An upgraded node answers `200` with `partial` instead.
        '503':
          description: The answering node could not read its own copy of the cluster metadata (for example while it catches up after a restart). Retry.
      x-narad-examples:
        - request: |
            curl -i -u "$AUTH" "$NARAD/v1/topics/orders?partition=1"
          response: |
            HTTP/1.1 200 OK
            Content-Length: 470
            Content-Type: application/json
            Date: Mon, 28 Sep 2026 19:31:43 GMT

            {"name":"orders","id":"681a429c0b683b2d","partitions":3,"retention_ms":172800000,"visibility_timeout_ms":30000,"max_in_flight_per_partition":1024,"max_acked_ahead_per_partition":1024,"created_at":1790623903,"owner":"admin","role":"parent","children":["orders-audit"],"schema_version":0,"partition_stats":[{"index":1,"segments":1,"oldest_offset":0,"next_offset":2,"high_watermark":2,"size_bytes":166,"oldest_segment_at":1790623903,"owner_node":"narad-0","status":"ok"}]}
    patch:
      operationId: alterTopic
      tags: [Topics]
      summary: Change a topic
      description: |
        Changes retention, the per-partition caps or the partition count, or
        registers a new schema version. Send at least one field.
        `visibility_timeout_ms` is fixed when the topic is created.

        Fields are applied one group at a time in this order: retention,
        then the caps, then partitions, then schema. There is no
        transaction across them: the first failure answers its error and
        the groups before it stay applied. Send one field per request when
        you need all or nothing.
      x-narad-grant: |
        Ownership of the topic, or `admin`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlterTopicRequest'
      responses:
        '200':
          description: Changed. The body is the topic after the change.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Topic'
        '400':
          description: No field given, a value is invalid, `partitions` is not larger than the current count, or the new schema is not compatible.
        '401':
          description: Missing or wrong credentials.
        '403':
          description: Not the owner and not `admin`, as the node that answers or the cluster leader sees the topic.
        '404':
          description: The topic does not exist, also after the answering node caught up with the leader.
        '409':
          description: '`schema_base_version` is not the current version, the history holds 1000 versions or (from v3.1.0) the new version would take it past 4 MiB or the cluster''s schemas past 256 MiB, the topic is a child whose schema its parent manages, the new retention is too short for a delay child, or (from v3.1.0) the topic was deleted and recreated, or grew, under the request twice in a row (`topic changed since it was read`). Unreleased: the topic is a remote child''s stub, which takes no change, or the new retention is below 24 hours on a parent with remote children (a retention already below it may still grow).'
        '413':
          description: The body is over 1 MiB.
        '415':
          description: No accepted `Content-Type` and no `X-Narad-Client` header.
        '503':
          description: The cluster has no leader to write the change, the answering node could not reach the leader to confirm a topic it does not have, or (from v3.1.0) a partition increase found every live node being decommissioned, so no node can take the new partitions; nothing was changed.
      x-narad-examples:
        - request: |
            curl -i -u "$AUTH" -X PATCH "$NARAD/v1/topics/orders" \
              -H "Content-Type: application/json" \
              -d '{"retention_ms": 172800000}'
          response: |
            HTTP/1.1 200 OK
            Content-Length: 244
            Content-Type: application/json
            Date: Mon, 28 Sep 2026 19:31:43 GMT

            {"name":"orders","id":"681a429c0b683b2d","partitions":3,"retention_ms":172800000,"visibility_timeout_ms":30000,"max_in_flight_per_partition":1024,"max_acked_ahead_per_partition":1024,"created_at":1790623903,"owner":"admin","role":"standalone"}
    delete:
      operationId: deleteTopic
      tags: [Topics]
      summary: Delete a topic
      description: |
        Deletes the topic's metadata and its data on every node, and
        removes its fan-out links. There is no undo. Once the metadata
        delete is committed the answer is `204`, even if a node that is
        down could not remove its files yet; that node removes them when
        it next starts.

        **Unreleased:** deleting a parent that has
        [remote children](glossary.md#remote-child) deletes their stubs in
        the same entry, and deleting a remote child's stub deletes the link.
        Either is refused with `409` while any record of the parent is not
        yet on the remote, unless `force=true` abandons them; the leader
        checks every member's ingress backlog and every cursor first. The
        same refusal, with its body, is in
        [Detach a child](#detach-child).
      x-narad-grant: |
        Ownership of the topic, or `admin`. Unreleased: a remote child's
        stub has no owner; the owner of its parent, or an `admin`, deletes
        it, with security on.
      parameters:
        - name: force
          in: query
          x-narad-since: unreleased
          description: |
            `true` deletes a parent with remote children, or a remote
            child's stub, even while records are unshipped, abandoning
            them. Ignored for any other topic. `narad topic rm --force`
            never sends it: there `--force` only skips the prompt.
          schema:
            type: boolean
            default: false
      responses:
        '204':
          description: Deleted.
        '400':
          description: Unreleased. `force` is not `true` or `false`.
        '401':
          description: Missing or wrong credentials.
        '403':
          description: Not the owner and not `admin`, as the node that answers or the cluster leader sees the topic. For a remote child's stub (unreleased), the caller is neither an `admin` nor the parent's owner, or security is off. For a parent with remote children (unreleased), security is off (`remotes require security`).
        '404':
          description: The topic does not exist, also after the answering node caught up with the leader.
        '409':
          description: New in v3.1.0. The topic was deleted and recreated under the request twice in a row (`topic changed since it was read`); nothing was deleted. Read the topic again before deleting it. Unreleased, for a parent with remote children or a stub without `force`, records of the parent are not yet on the remote (the body carries `lag_messages`, `lag_complete` and `dispatch_backlog`), or the leader found a remote child the answering node did not know of yet; retry.
        '412':
          description: Unreleased. The cluster leader runs a release without remote children; finish the upgrade.
        '429':
          description: Unreleased. The leader ran an unshipped check for this parent less than 10 seconds ago; retry after `Retry-After`.
          x-narad-origin: forwarding
        '501':
          description: Unreleased. The answering node has no remote plane, so it cannot delete a remote-linked topic.
        '503':
          description: The cluster has no leader to write the delete, or the answering node could not reach the leader to confirm a topic it does not have. Unreleased, for a remote-linked topic, also when the unshipped check could not run.
      x-narad-examples:
        - request: |
            curl -i -u "$AUTH" -X DELETE "$NARAD/v1/topics/orders-audit"
          response: |
            HTTP/1.1 204 No Content
            Date: Mon, 28 Sep 2026 19:31:43 GMT
...
  /v1/topics/{parent}/children/{child}:
    parameters:
      - $ref: '#/components/parameters/ParentPath'
      - name: child
        in: path
        required: true
        description: Name of the child topic.
        schema:
          type: string
    delete:
      operationId: detachChild
      tags: [Fan-out children]
      summary: Detach a child
      description: |
        Removes the link. The child keeps the messages and the schema
        history it already has and becomes a standalone topic again.

        **Unreleased:** detaching a [remote child](glossary.md#remote-child)
        deletes its stub. It travels to the leader, which first checks that
        every record of the parent is on the remote: no cursor lag, every
        partition owner reporting, and no member holding records of the
        parent it answered `202` for and has not committed yet. While any
        is left it answers `409` with the counts, unless `force=true`
        abandons them; the leader's audit line records what a forced delete
        abandoned. The check is a point in time, so stop the producers
        first. A new check for one parent runs at most every 10 seconds.
      x-narad-grant: |
        Ownership of either topic, or `admin`. Unreleased: for a remote
        child, the owner of the parent, or an `admin`, with security on.
      parameters:
        - name: force
          in: query
          x-narad-since: unreleased
          description: '`true` deletes a remote child even while records of its parent are unshipped, abandoning them. Ignored for a local child.'
          schema:
            type: boolean
            default: false
      responses:
        '204':
          description: Detached. For a remote child (unreleased), its stub is deleted.
        '400':
          description: Unreleased. `force` is not `true` or `false`.
        '401':
          description: Missing or wrong credentials.
        '403':
          description: The caller manages neither topic, as the node that answers or the cluster leader sees them. For a remote child (unreleased), the caller is neither an `admin` nor the parent's owner, or security is off.
        '404':
          description: Neither topic exists after the answering node caught up with the leader, or the two are not linked.
        '409':
          description: 'New in v3.1.0. Either topic was deleted and recreated under the request twice in a row (`topic changed since it was read`); nothing was changed. Unreleased, for a remote child without `force`: records of the parent are not yet on the remote, with `lag_messages`, `lag_complete` and `dispatch_backlog` (records each member holds, by node) in the body, and `not_answering` or `backlog_over_scan_limit` naming members that could not be counted; or the child turned out to be remote, or local, on the leader; retry.'
        '412':
          description: Unreleased. The cluster leader runs a release without remote children; finish the upgrade.
        '429':
          description: Unreleased. The leader ran an unshipped check for this parent less than 10 seconds ago; retry after `Retry-After`.
          x-narad-origin: forwarding
        '501':
          description: Unreleased. The answering node has no remote plane, so it cannot detach a remote child.
        '503':
          description: The cluster has no leader to write the change, or the answering node could not reach the leader to confirm the topics. Unreleased, for a remote child, also when the unshipped check could not run.
      x-narad-examples:
        - title: a local child
          request: |
            curl -i -u "$AUTH" -X DELETE \
              "$NARAD/v1/topics/orders/children/orders-audit"
          response: |
            HTTP/1.1 204 No Content
            Date: Mon, 28 Sep 2026 19:31:43 GMT
        - title: a remote child with records not yet on its remote (unreleased)
          request: |
            curl -i -u "$AUTH" -X DELETE \
              "$NARAD/v1/topics/orders/children/orders-to-b"
          response: |
            HTTP/1.1 409 Conflict
            Cache-Control: no-store
            Content-Type: application/json
            X-Content-Type-Options: nosniff
            Date: Tue, 06 Oct 2026 13:07:18 GMT
            Content-Length: 122

            {"dispatch_backlog":{},"error":"remote child \"orders-to-b\" has unshipped records","lag_complete":true,"lag_messages":4}
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  parameters:
    TopicPath:
      name: topic
      in: path
      required: true
      description: Name of the topic.
      schema:
        type: string
    ParentPath:
      name: parent
      in: path
      required: true
      description: Name of the parent topic.
      schema:
        type: string
...
  schemas:
    CreateTopicRequest:
      type: object
      required: [name]
      additionalProperties: false
      properties:
        name:
          type: string
          maxLength: 200
          description: |
            1 to 200 characters from `A-Z a-z 0-9 . _ -`, not `.` or `..`,
            and not differing from an existing topic's name only in letter
            case (`409`). Topics an earlier release created with names up to
            255 characters keep working.
        partitions:
          type: integer
          minimum: 3
          description: |
            At least 3 and at most the server's maximum (108 by default).
            Defaults to the server default (3), or to the parent's count
            when `parent` is set. It can grow later, never shrink. Both
            server settings are in the
            [Configuration reference](configuration.md#topic-defaults).
        retention_ms:
          type: integer
          description: |
            How long a message is kept after it is written, whether or not
            it was consumed. At least 3,600,000 (1 hour), or `0` to keep
            messages forever; a negative value gets `400`. Left out, it
            takes the server default
            ([Configuration reference](configuration.md#topic-defaults)):
            7 days in the binary, 12 hours in the Helm chart.
        visibility_timeout_ms:
          type: integer
          default: 30000
          description: How long a consumer's lease lasts before the message is delivered again. Defaults to the server default ([Configuration reference](configuration.md#topic-defaults)). Fixed after creation.
        max_in_flight_per_partition:
          type: integer
          default: 1024
          description: Most messages of one partition leased at once. At the cap, consumes find nothing new in that partition until a lease ends. Defaults to the server default ([Configuration reference](configuration.md#topic-defaults)).
        max_acked_ahead_per_partition:
          type: integer
          default: 1024
          description: |
            Most acks one partition holds above its oldest unacked message.
            At the cap the partition delivers no new messages until that
            message is acked. Defaults to the server default
            ([Configuration reference](configuration.md#topic-defaults)).
        schema:
          description: A JSON Schema every message must match; registered as version 1. A JSON object or `true`.
        parent:
          type: string
          description: |
            Create the topic as a fan-out child of this existing topic. Its
            partitions are placed on other nodes than the parent's where the
            cluster allows, which is what makes a
            [replica child](../operate/backups.md#replica-children).
        fanout_delay_ms:
          type: integer
          description: |
            With `parent`, make the topic a delay child that receives each
            message this long after the parent committed it. At most
            31,536,000,000 (one year). Attaching an existing topic takes
            `delay_ms` instead.
        owner:
          type: string
          description: Ignored. The server sets the owner to the caller.
...
    Topic:
      type: object
      properties:
        name:
          type: string
          description: Topic name.
        id:
          type: string
          description: The [incarnation](glossary.md#incarnation) ID, 16 hex characters. A topic deleted and created again under the same name gets a new one. Absent on a topic created before v2.2.0.
        partitions:
          type: integer
          description: Partition count.
        retention_ms:
          type: integer
          description: Retention in milliseconds; `0` keeps messages forever.
        visibility_timeout_ms:
          type: integer
          description: Lease length in milliseconds.
        max_in_flight_per_partition:
          type: integer
          description: In-flight cap per partition.
        max_acked_ahead_per_partition:
          type: integer
          description: Acked-ahead cap per partition.
        created_at:
          type: integer
          description: Creation time, Unix seconds.
        owner:
          type: string
          description: The user that created the topic. Absent when security was off.
        role:
          type: string
          enum: [standalone, parent, child]
          description: Fan-out role. Absent in the answer to a create without `parent`, which means `standalone`.
        children:
          type: array
          items:
            type: string
          description: A parent's children, in attach order.
        parent:
          type: string
          description: A child's parent.
        attach_epoch:
          type: string
          description: A child's current attachment; it changes on every attach.
        fanout_delay_ms:
          type: integer
          description: A delay child's delay in milliseconds.
        attach_offsets:
          type: array
          items:
            type: integer
          description: A child's attach point, one offset per parent partition.
        remote:
          $ref: '#/components/schemas/RemoteLink'
...
    TopicPage:
      type: object
      properties:
        topics:
          type: array
          description: Topics the caller can read.
          items:
            $ref: '#/components/schemas/Topic'
        next_page_token:
          type: string
          description: Pass as `page_token` for the next page; empty when there are no more.
...